pyproject.toml
[tool.poetry]
name = "leet_chat_backend"
version = "0.1.0"
description = ""
authors = ["dhvcc <1337kwiz@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
django-rest-framework = "^0.1.0"
psycopg2-binary = "^2.8.6"
PyYAML = "^5.3.1"
uritemplate = "^3.0.1"
coreapi = "^2.3.3"
django-baton = "^1.12.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pytest = "^6.1.2"
Werkzeug = "0.16.1"
django-extensions = "^3.0.9"
python-dotenv = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
exclude = "migrations\\/"